# contracts/openapi/mvp-v0.2.json
{
  "components": {
    "schemas": {
...
      "AdvanceRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "AdvanceRequest",
        "type": "object"
      },
      "AdvanceResult": {
        "additionalProperties": false,
        "properties": {
          "match": {
            "$ref": "#/components/schemas/MatchSnapshot"
          },
          "round": {
            "$ref": "#/components/schemas/RoundExecutionPublicView"
          },
          "schema_version": {
            "const": "mvp-v0.2",
            "default": "mvp-v0.2",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "round",
          "match"
        ],
        "title": "AdvanceResult",
        "type": "object"
      },
      "BattleEscalationSnapshot": {
        "additionalProperties": false,
        "properties": {
          "hard_liveness_active": {
            "title": "Hard Liveness Active",
            "type": "boolean"
          },
          "level": {
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Level",
            "type": "integer"
          },
          "next_level_at_no_damage": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Level At No Damage"
          },
          "no_damage_streak": {
            "minimum": 0.0,
            "title": "No Damage Streak",
            "type": "integer"
          },
          "rounds_until_next_level": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds Until Next Level"
          }
        },
        "required": [
          "level",
          "no_damage_streak",
          "hard_liveness_active"
        ],
        "title": "BattleEscalationSnapshot",
        "type": "object"
      },
      "BoardSnapshot": {
        "additionalProperties": false,
        "properties": {
          "cols": {
            "minimum": 1.0,
            "title": "Cols",
            "type": "integer"
          },
          "rows": {
            "minimum": 1.0,
            "title": "Rows",
            "type": "integer"
          }
        },
        "required": [
          "rows",
          "cols"
        ],
        "title": "BoardSnapshot",
        "type": "object"
      },
      "CreateMatchRequest": {
        "additionalProperties": false,
        "properties": {
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "title": "CreateMatchRequest",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "schema_version": {
            "const": "mvp-v0.2",
            "default": "mvp-v0.2",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchLifecycle": {
        "enum": [
          "RUNNING",
          "PLAYER_DECISION",
          "TERMINAL",
          "FAILED_RECOVERABLE"
        ],
        "title": "MatchLifecycle",
        "type": "string"
      },
      "MatchResultPublic": {
        "enum": [
          "RED_WIN",
          "BLUE_WIN",
          "DRAW_MUTUAL_DEATH",
          "TIMEOUT"
        ],
        "title": "MatchResultPublic",
        "type": "string"
      },
      "MatchSnapshot": {
        "additionalProperties": false,
        "properties": {
          "active_rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RulePublicView"
              },
              {
                "type": "null"
              }
            ]
          },
          "battle_escalation": {
            "$ref": "#/components/schemas/BattleEscalationSnapshot"
          },
          "board": {
            "$ref": "#/components/schemas/BoardSnapshot"
          },
          "completed_rounds": {
            "minimum": 0.0,
            "title": "Completed Rounds",
            "type": "integer"
          },
          "effective_stats": {
            "$ref": "#/components/schemas/TeamStatsMap"
          },
          "latest_strategy": {
            "$ref": "#/components/schemas/TeamLatestStrategyMap"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/MatchLifecycle"
          },
          "match_id": {
            "minLength": 1,
            "title": "Match Id",
            "type": "string"
          },
          "player_decision": {
            "$ref": "#/components/schemas/PlayerDecisionSnapshot"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchResultPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "round_no": {
            "minimum": 1.0,
            "title": "Round No",
            "type": "integer"
          },
          "rule_change_count": {
            "minimum": 0.0,
            "title": "Rule Change Count",
            "type": "integer"
          },
          "schema_version": {
            "const": "mvp-v0.2",
            "default": "mvp-v0.2",
            "title": "Schema Version",
            "type": "string"
          },
          "score_rounds": {
            "minimum": 0.0,
            "title": "Score Rounds",
            "type": "integer"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "units": {
            "$ref": "#/components/schemas/TeamUnitMap"
          }
        },
        "required": [
          "match_id",
          "revision",
          "lifecycle",
          "seed",
          "round_no",
          "completed_rounds",
          "score_rounds",
          "rule_change_count",
          "board",
          "units",
          "active_rule",
          "player_decision",
          "effective_stats",
          "latest_strategy",
          "battle_escalation",
          "result"
        ],
        "title": "MatchSnapshot",
        "type": "object"
      },
...
      "ModelCallLogExport": {
        "additionalProperties": false,
        "properties": {
          "attempted_calls": {
            "minimum": 0.0,
            "title": "Attempted Calls",
            "type": "integer"
          },
          "confirmed_responses": {
            "minimum": 0.0,
            "title": "Confirmed Responses",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ModelCallEntryPublic"
            },
            "title": "Entries",
            "type": "array"
          },
          "failed_attempts": {
            "minimum": 0.0,
            "title": "Failed Attempts",
            "type": "integer"
          },
          "log_version": {
            "const": "model-call-log-v1",
            "default": "model-call-log-v1",
            "title": "Log Version",
            "type": "string"
          },
          "match_id": {
            "minLength": 1,
            "title": "Match Id",
            "type": "string"
          },
          "retained_entries": {
            "minimum": 0.0,
            "title": "Retained Entries",
            "type": "integer"
          },
          "schema_version": {
            "const": "mvp-v0.2",
            "default": "mvp-v0.2",
            "title": "Schema Version",
            "type": "string"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "match_id",
          "attempted_calls",
          "confirmed_responses",
          "failed_attempts",
          "retained_entries",
          "truncated",
          "entries"
        ],
        "title": "ModelCallLogExport",
        "type": "object"
      },
...
      "PlayerDecisionSnapshot": {
        "additionalProperties": false,
        "properties": {
          "after_round": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Round"
          },
          "can_advance": {
            "title": "Can Advance",
            "type": "boolean"
          },
          "can_submit_rule": {
            "title": "Can Submit Rule",
            "type": "boolean"
          },
          "rule_changed_this_intermission": {
            "title": "Rule Changed This Intermission",
            "type": "boolean"
          }
        },
        "required": [
          "can_submit_rule",
          "rule_changed_this_intermission",
          "can_advance"
        ],
        "title": "PlayerDecisionSnapshot",
        "type": "object"
      },
...
      "ReplaySnapshot": {
        "additionalProperties": false,
        "properties": {
          "initial_config": {
            "$ref": "#/components/schemas/GameConfigPublicView"
          },
          "match_id": {
            "minLength": 1,
            "title": "Match Id",
            "type": "string"
          },
          "replay_version": {
            "const": "replay-v0.2",
            "default": "replay-v0.2",
            "title": "Replay Version",
            "type": "string"
          },
          "rule_change_count": {
            "minimum": 0.0,
            "title": "Rule Change Count",
            "type": "integer"
          },
          "schema_version": {
            "const": "mvp-v0.2",
            "default": "mvp-v0.2",
            "title": "Schema Version",
            "type": "string"
          },
          "score_rounds": {
            "minimum": 0.0,
            "title": "Score Rounds",
            "type": "integer"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "terminal_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchResultPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeline": {
            "items": {
              "discriminator": {
                "mapping": {
                  "INTERMISSION": "#/components/schemas/ReplayIntermissionEntry",
                  "ROUND": "#/components/schemas/ReplayRoundEntry"
                },
                "propertyName": "entry_type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ReplayIntermissionEntry"
                },
                {
                  "$ref": "#/components/schemas/ReplayRoundEntry"
                }
              ]
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "match_id",
          "seed",
          "initial_config",
          "timeline",
          "terminal_result",
          "score_rounds",
          "rule_change_count"
        ],
        "title": "ReplaySnapshot",
        "type": "object"
      },
...
      "RoundExecutionPublicView": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "$ref": "#/components/schemas/TeamActionMap"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RoundEventPublicView"
            },
            "title": "Events",
            "type": "array"
          },
          "round_no": {
            "minimum": 1.0,
            "title": "Round No",
            "type": "integer"
          },
          "strategies": {
            "$ref": "#/components/schemas/TeamRoundStrategyMap"
          }
        },
        "required": [
          "round_no",
          "strategies",
          "actions",
          "events"
        ],
        "title": "RoundExecutionPublicView",
        "type": "object"
      },
      "RuleAstPublicView": {
        "additionalProperties": false,
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/RuleConditionPublicView"
            },
            "title": "Conditions",
            "type": "array"
          },
          "duration": {
            "$ref": "#/components/schemas/RuleDurationPublic"
          },
          "effect": {
            "$ref": "#/components/schemas/RuleEffectPublicView"
          },
          "target": {
            "$ref": "#/components/schemas/RuleTargetPublic"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "target",
          "conditions",
          "effect",
          "duration"
        ],
        "title": "RuleAstPublicView",
        "type": "object"
      },
...
      "RuleDurationPublic": {
        "enum": [
          "UNTIL_REPLACED"
        ],
        "title": "RuleDurationPublic",
        "type": "string"
      },
...
      "RulePublicView": {
        "additionalProperties": false,
        "properties": {
          "ast": {
            "$ref": "#/components/schemas/RuleAstPublicView"
          },
          "player_text": {
            "title": "Player Text",
            "type": "string"
          },
          "rule_id": {
            "minLength": 1,
            "title": "Rule Id",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "player_text",
          "ast"
        ],
        "title": "RulePublicView",
        "type": "object"
      },
...
      "RuleSubmissionRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "player_text": {
            "minLength": 1,
            "title": "Player Text",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "player_text"
        ],
        "title": "RuleSubmissionRequest",
        "type": "object"
      },
      "RuleSubmissionResult": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "candidate_preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleAstPublicView"
              },
              {
                "type": "null"
              }
            ]
          },
          "match": {
            "$ref": "#/components/schemas/MatchSnapshot"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "public_code": {
            "$ref": "#/components/schemas/RuleSubmissionCode"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "schema_version": {
            "const": "mvp-v0.2",
            "default": "mvp-v0.2",
            "title": "Schema Version",
            "type": "string"
          },
          "suggested_rephrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Rephrase"
          }
        },
        "required": [
          "accepted",
          "public_code",
          "message",
          "match"
        ],
        "title": "RuleSubmissionResult",
        "type": "object"
      },
...
      "TeamLatestStrategyMap": {
        "additionalProperties": false,
        "properties": {
          "BLUE": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicStrategyDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "RED": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicStrategyDecision"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "RED",
          "BLUE"
        ],
        "title": "TeamLatestStrategyMap",
        "type": "object"
      },
...
      "TeamStatsMap": {
        "additionalProperties": false,
        "properties": {
          "BLUE": {
            "$ref": "#/components/schemas/EffectiveStatsPublicView"
          },
          "RED": {
            "$ref": "#/components/schemas/EffectiveStatsPublicView"
          }
        },
        "required": [
          "RED",
          "BLUE"
        ],
        "title": "TeamStatsMap",
        "type": "object"
      },
      "TeamUnitMap": {
        "additionalProperties": false,
        "properties": {
          "BLUE": {
            "$ref": "#/components/schemas/UnitSnapshot"
          },
          "RED": {
            "$ref": "#/components/schemas/UnitSnapshot"
          }
        },
        "required": [
          "RED",
          "BLUE"
        ],
        "title": "TeamUnitMap",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Frozen MVP V0.2 public HTTP contract for the every-round player-decision flow. Route bodies are implemented by MatchApplicationService later.",
    "title": "Rules Beyond MVP API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/matches": {
      "post": {
        "operationId": "create_match_api_v1_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Match"
      }
    },
    "/api/v1/matches/{match_id}": {
      "get": {
        "operationId": "get_match_api_v1_matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match"
      }
    },
    "/api/v1/matches/{match_id}/advance": {
      "post": {
        "operationId": "advance_match_api_v1_matches__match_id__advance_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Advance Match"
      }
    },
    "/api/v1/matches/{match_id}/model-calls": {
      "get": {
        "operationId": "get_model_calls_api_v1_matches__match_id__model_calls_get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCallLogExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model Calls"
      }
    },
    "/api/v1/matches/{match_id}/replay": {
      "get": {
        "operationId": "get_replay_api_v1_matches__match_id__replay_get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaySnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Replay"
      }
    },
    "/api/v1/matches/{match_id}/rules": {
      "post": {
        "operationId": "submit_rule_api_v1_matches__match_id__rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSubmissionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Rule"
      }
    }
  }
}